perubahan semua router link master dan transaksi di semua pages

+++ projects/iam/src/app/pages/companies/companies-list/companies-list.component.ts
@@ -1,15 +1,18 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { ConfirmationService } from 'primeng/api';
 import { FindAllResCompaniesDto } from 'projects/core/src/app/dto/companies/find-all-res-companies-dto';
 import { Companies } from 'projects/core/src/app/model/companies';
 import { AuthenticationService } from 'projects/core/src/app/services/authentication/authentication.service';
 import { CompaniesService } from 'projects/core/src/app/services/companies/companies.service';
 import { Subscription } from 'rxjs';
 
+
 @Component({
   selector: 'app-companies-list',
   templateUrl: './companies-list.component.html',
-  styleUrls: ['./companies-list.component.css']
+  styleUrls: ['./companies-list.component.css'],
+  providers: [ConfirmationService]
 })
 export class CompaniesListComponent implements OnInit, OnDestroy {
 
@@ -20,13 +23,13 @@ export class CompaniesListComponent implements OnInit, OnDestroy {
   listCompany: Companies[] = []
 
   constructor(private router: Router, private companiesService: CompaniesService,
-    private authService: AuthenticationService) { }
-
-  ngOnDestroy(): void {
-    this.obs?.unsubscribe()
-  }
+    private authService: AuthenticationService, private confirmationService: ConfirmationService) { }
 
   ngOnInit(): void {
+    this.initData()
+  }
+
+  initData(): void {
     this.allDataCompanies = new FindAllResCompaniesDto()
     this.companiesService.findAllCompanies().subscribe(result => {
       this.allDataCompanies = result
@@ -35,19 +38,27 @@ export class CompaniesListComponent implements OnInit, OnDestroy {
   }
 
   clickCreate(): void {
-    this.router.navigateByUrl('/companies-action/new')
+    this.router.navigateByUrl('/companies/new')
   }
 
   clickUpdate(id: string): void {
-    this.router.navigateByUrl(`/companies-action/${id}`)
+    this.router.navigateByUrl(`/companies/modify/${id}`)
   }
 
-  clickDelete(id: string): void {
-    this.companiesService.delete(id).subscribe({
-      next: result => {
-        window.location.reload()
+  confirm(id: string): void {
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to Remove?',
+      accept: () => {
+        this.companiesService.delete(id).subscribe({
+          next: result => {
+            this.initData()
+          }
+        })
       }
-    })
+    });
   }
 
+  ngOnDestroy(): void {
+    this.obs?.unsubscribe()
+  }
 }
