+++ projects/iam/src/app/pages/employees/employees-list/employees-list.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { ConfirmationService } from 'primeng/api';
 import { FindAllResEmployeesDto } from 'projects/core/src/app/dto/employees/find-all-res-employees-dto';
 import { Employees } from 'projects/core/src/app/model/employees';
 import { AuthenticationService } from 'projects/core/src/app/services/authentication/authentication.service';
@@ -9,7 +10,8 @@ import { Subscription } from 'rxjs';
 @Component({
   selector: 'app-employees-list',
   templateUrl: './employees-list.component.html',
-  styleUrls: ['./employees-list.component.css']
+  styleUrls: ['./employees-list.component.css'],
+  providers: [ConfirmationService]
 })
 export class EmployeesListComponent implements OnInit, OnDestroy {
 
@@ -19,14 +21,18 @@ export class EmployeesListComponent implements OnInit, OnDestroy {
 
   listEmployee: Employees[] = []
 
-  constructor(private router: Router, private employeesService: EmployeesService,
-    private authService: AuthenticationService) { }
+  constructor(private router: Router, private authService: AuthenticationService, 
+    private employeesService: EmployeesService, private confirmationService: ConfirmationService) { }
 
   ngOnDestroy(): void {
     this.obs?.unsubscribe()
   }
 
   ngOnInit(): void {
+    this.initData()
+  }
+
+  initData(): void {
     this.allDataEmployees = new FindAllResEmployeesDto()
     this.employeesService.findAllEmployees().subscribe(result => {
       this.allDataEmployees = result
@@ -35,19 +41,24 @@ export class EmployeesListComponent implements OnInit, OnDestroy {
   }
 
   clickCreate(): void {
-    this.router.navigateByUrl('/employees-action/new')
+    this.router.navigateByUrl('/employees/new')
   }
 
   clickUpdate(id: string): void {
-    this.router.navigateByUrl(`/employees-action/${id}`)
+    this.router.navigateByUrl(`/employees/modify/${id}`)
   }
 
-  clickDelete(id: string): void {
-    this.employeesService.delete(id).subscribe({
-      next: result => {
-        window.location.reload()
+  confirm(id: string): void {
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to Remove?',
+      accept: () => {
+        this.employeesService.delete(id).subscribe({
+          next: result => {
+            this.initData()
+          }
+        })
       }
-    })
+    });
   }
 
   clickBack() {
@@ -55,15 +66,3 @@ export class EmployeesListComponent implements OnInit, OnDestroy {
   }
 
 }
-
-// class Employees {
-//   number?: number
-//   company?: string
-//   nip?: string
-//   fullName?: string
-//   email?: string
-//   phoneNo?: string
-//   department?: string
-//   isActive?: boolean
-// }
-
