+++ projects/iam/src/app/pages/locations/locations-list/locations-list.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { ConfirmationService } from 'primeng/api';
 import { FindAllResLocationsDto } from 'projects/core/src/app/dto/locations/find-all-res-locations-dto';
 import { Locations } from 'projects/core/src/app/model/locations';
 import { LocationsService } from 'projects/core/src/app/services/locations/locations.service';
@@ -7,13 +8,21 @@ import { LocationsService } from 'projects/core/src/app/services/locations/locat
 @Component({
   selector: 'app-locations-list',
   templateUrl: './locations-list.component.html',
-  styleUrls: ['./locations-list.component.css']
+  styleUrls: ['./locations-list.component.css'],
+  providers: [ConfirmationService]
 })
 export class LocationsListComponent implements OnInit {
+
   listLocations : Locations[] = []
-  constructor(private locationService : LocationsService, private router : Router) { }
+
+  constructor(private router : Router, private locationService : LocationsService,
+    private confirmationService: ConfirmationService) { }
 
   ngOnInit(): void {
+    this.initData()
+  }
+
+  initData(): void {
     this.locationService.findAll().subscribe({
       next : result=>{
         const locationResult : FindAllResLocationsDto = result
@@ -23,17 +32,24 @@ export class LocationsListComponent implements OnInit {
   }
 
   clickCreate():void{
-    this.router.navigateByUrl('location-action/new')
+    this.router.navigateByUrl('locations/new')
   }
+
   clickUpdate(id : string){
-    this.router.navigateByUrl(`location-action/${id}`)
+    this.router.navigateByUrl(`locations/modify/${id}`)
   }
-  clickDelete(id : string){
-    this.locationService.delete(id).subscribe(
-      result =>{
-        window.location.reload()
+
+   confirm(id: string): void {
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to Remove?',
+      accept: () => {
+        this.locationService.delete(id).subscribe({
+          next: result => {
+            this.initData()
+          }
+        })
       }
-    )
+    });
   }
 
 }
